requirements.txt
streamlit>=1.36.0
anthropic>=0.34.0
python-docx>=1.1.0
mammoth>=1.8.0
